simplfy part of the code
